Done - prepopulating update form .html
also bug fixed.

+++ OnlineSoccer/src/main/java/com/example/entity/Book.java
@@ -38,7 +38,7 @@ public class Book {
 	public Book() {
 	}
 
-	public float getId() {
+	public int getId() {
 		return id;
 	}
